Expose htable
