+++ src/hash.c
@@ -21,32 +21,6 @@ mrb_int mrb_float_id(mrb_float f);
 #endif
 #define HT_SEG_INCREASE_RATIO 6 / 5
 
-struct segkv {
-  mrb_value key;
-  mrb_value val;
-};
-
-typedef struct segment {
-  uint16_t size;
-  struct segment *next;
-  struct segkv e[];
-} segment;
-
-typedef struct segindex {
-  size_t size;
-  size_t capa;
-  struct segkv *table[];
-} segindex;
-
-/* hash table structure */
-typedef struct htable {
-  segment *rootseg;
-  segment *lastseg;
-  mrb_int size;
-  uint16_t last_len;
-  segindex *index;
-} htable;
-
 static /* inline */ size_t
 ht_hash_func(mrb_state *mrb, htable *t, mrb_value key)
 {
